pyproject.toml
[project]
name = "pinyin_subs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.64.0",
    "pysrt>=1.1.2",
    "python-dotenv>=1.0.1",
]
